+++ __init__.py
@@ -17,7 +17,7 @@ if sys.version_info < (3, 10):
 import supybot
 from supybot import world
 
-__version__ = "1.1.0-beta.1"
+__version__ = "1.1.0-beta.2"
 __author__ = supybot.Author("Barry Suridge", "Alcheri", "barry.suridge@gmail.com")
 __contributors__ = {}
 __url__ = "https://github.com/Alcheri/Geminoria "